requirements.txt
Flask
pandas
openai
python-dotenv